+++ TrippyTesting/Tests/Test3DBatcher.cs
@@ -442,7 +442,7 @@ namespace TrippyTesting.Tests
             if (size.Width == 0 || size.Height == 0)
                 return;
 
-            graphicsDevice.SetViewport(0, 0, size.Width, size.Height);
+            graphicsDevice.SetViewport(0, 0, (uint)size.Width, (uint)size.Height);
 
             float wid = size.Width / (float)size.Height;
             wid *= 0.5f;
